opt: move sql/opt/opt files to sql/opt
Release note: None
