+++ pkg/sql/opt/metadata.go
