+++ pkg/sql/opt/metadata_test.go
