+++ pkg/sql/opt/optbuilder/builder.go
@@ -18,7 +18,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 )
