+++ pkg/sql/opt/optgen/cmd/optgen/exprs_gen.go
@@ -35,7 +35,7 @@ func (g *exprsGen) generate(compiled *lang.CompiledExpr, w io.Writer) {
 	fmt.Fprintf(g.w, "package xform\n\n")
 
 	fmt.Fprintf(g.w, "import (\n")
-	fmt.Fprintf(g.w, "  \"github.com/cockroachdb/cockroach/pkg/sql/opt/opt\"\n")
+	fmt.Fprintf(g.w, "  \"github.com/cockroachdb/cockroach/pkg/sql/opt\"\n")
 	fmt.Fprintf(g.w, ")\n\n")
 
 	g.genChildCountLookup()
