+++ pkg/sql/opt/optgen/cmd/optgen/testdata/factory
@@ -26,7 +26,7 @@ define Sort {
 package xform
 
 import (
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 )
 
 // ConstructNot constructs an expression for the Not operator.
@@ -112,7 +112,7 @@ define InnerJoin {
 package xform
 
 import (
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 )
 
 // ConstructInnerJoin constructs an expression for the InnerJoin operator.
@@ -192,7 +192,7 @@ define Ne {
 package xform
 
 import (
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 )
 
 // ConstructEq constructs an expression for the Eq operator.
@@ -304,7 +304,7 @@ define Variable {
 package xform
 
 import (
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 )
 
 // ConstructFunc constructs an expression for the Func operator.
@@ -435,7 +435,7 @@ define And {
 package xform
 
 import (
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 )
 
 // ConstructSelect constructs an expression for the Select operator.
@@ -631,7 +631,7 @@ define List {
 package xform
 
 import (
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 )
 
 // ConstructList constructs an expression for the List operator.
@@ -748,7 +748,7 @@ define Ne {
 package xform
 
 import (
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 )
 
 // ConstructEq constructs an expression for the Eq operator.
