+++ pkg/sql/opt/physical_props.go
