+++ pkg/sql/opt/physical_props_test.go
