+++ pkg/sql/opt/xform/expr.og.go
@@ -3,7 +3,7 @@
 package xform
 
 import (
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 )
 
 type childCountLookupFunc func(ev ExprView) int
