+++ pkg/sql/opt/xform/list_storage_test.go
@@ -18,7 +18,7 @@ import (
 	"bytes"
 	"testing"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 )
 
 func TestListStorage(t *testing.T) {
