+++ pkg/sql/opt/xform/logical_props_test.go
@@ -17,7 +17,7 @@ package xform
 import (
 	"testing"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/testutils"
 	"github.com/cockroachdb/cockroach/pkg/util/treeprinter"
 )
