+++ pkg/sql/opt/xform/memo_group.go
@@ -15,7 +15,7 @@
 package xform
 
 import (
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 	"github.com/cockroachdb/cockroach/pkg/util"
 )
