+++ pkg/sql/opt/xform/optimizer.go
@@ -17,7 +17,7 @@ package xform
 import (
 	"fmt"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/opt"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt"
 	"github.com/cockroachdb/cockroach/pkg/sql/optbase"
 	"golang.org/x/tools/container/intsets"
 )
